+++ src/AuthLDAP.php
@@ -2085,7 +2085,6 @@ TWIG, $twig_params);
     /**
      * Show LDAP groups to add or synchronize in an entity
      *
-     * @param string  $target  target page for the form
      * @param integer $start   where to start the list
      * @param integer $sync    synchronize or add? (default 0)
      * @param string  $filter  ldap filter to use (default '')
@@ -2094,10 +2093,9 @@ TWIG, $twig_params);
      *
      * @return void
      *
-     * @since 11.0.0 $order parameter has been removed.
+     * @since 11.0.0 The `$target` and the `$order` parameters have been removed.
      */
     public static function showLdapGroups(
-        $target,
         $start,
         $sync = 0,
         $filter = '',